requirements.txt
# Основные вычисления (совместимы с CUDA 11.8/12.1+)
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.3

scikit-learn>=1.3.0
scipy>=1.11.0

Pillow>=10.0.0
matplotlib>=3.7.0
pycocotools>=2.0.7

tqdm>=4.66.0